requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.7.0
certifi==2022.9.14
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.1
cmake==3.24.1.1
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
Django==4.1.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
exceptiongroup==1.0.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
iniconfig==1.1.1
Markdown==3.4.1
msgpack==1.0.4
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pluggy==1.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytest==7.2.0
pytest-asyncio==0.20.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.3.0
urllib3==1.26.12
zope.interface==5.4.0
